# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  library-management-system-go_internal_dto.CreateArticleCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  library-management-system-go_internal_dto.CreateArticleRequest:
    properties:
      category:
        type: string
      content:
        type: string
      excerpt:
        type: string
      featured_image_url:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - category
    - content
    - title
    type: object
...
  library-management-system-go_internal_dto.UpdateArticleCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  library-management-system-go_internal_dto.UpdateArticleRequest:
    properties:
      category:
        type: string
      content:
        type: string
      excerpt:
        type: string
      featured_image_url:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for school library management.
  title: School Library Management API
  version: "1.0"
paths:
  /admin/hello:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Admin hello
      tags:
      - access
  /articles:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      parameters:
      - description: Create article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - articles
  /articles/{id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete article
      tags:
      - articles
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get article
      tags:
      - articles
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - articles
  /articles/{id}/comments:
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List article comments
      tags:
      - article-comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create article comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article comment
      tags:
      - article-comments
  /articles/{id}/comments/{comment_id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete article comment
      tags:
      - article-comments
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: integer
      - description: Update article comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article comment
      tags:
      - article-comments
  /articles/{id}/ratings:
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List article ratings
      tags:
      - article-ratings
    post:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create article rating request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleRatingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article rating
      tags:
      - article-ratings
  /articles/{id}/ratings/{rating_id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete article rating
      tags:
      - article-ratings
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: integer
      - description: Update article rating request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article rating
      tags:
      - article-ratings
  /articles/review:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List article reviews
      tags:
      - article-reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Create article review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article review
      tags:
      - article-reviews
  /articles/review/{id}:
    get:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get article review
      tags:
      - article-reviews
    put:
      consumes:
      - application/json
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update article review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article review
      tags:
      - article-reviews
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Change password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library-management-system-go_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library-management-system-go_pkg_response.Response'
      summary: Login user
      tags:
      - auth
  /auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library-management-system-go_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/library-management-system-go_pkg_response.Response'
      summary: Register a new user
      tags:
      - auth
  /books:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List books
      tags:
      - books
    post:
      consumes:
      - application/json
      parameters:
      - description: Create book request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create book
      tags:
      - books
  /books/{id}:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete book
      tags:
      - books
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get book
      tags:
      - books
    put:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update book request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update book
      tags:
      - books
  /books/{id}/comments:
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List book comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create book comment
      tags:
      - comments
...
  /books/{id}/reviews:
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List book reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create book review
      tags:
      - reviews
...
  /complaints:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List complaints
      tags:
      - complaints
    post:
      consumes:
      - application/json
      parameters:
      - description: Create complaint request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateComplaintRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create complaint
      tags:
      - complaints
  /complaints/{id}:
    delete:
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete complaint
      tags:
      - complaints
    get:
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get complaint
      tags:
      - complaints
    put:
      consumes:
      - application/json
      parameters:
      - description: Complaint ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update complaint request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateComplaintRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update complaint
      tags:
      - complaints
  /fines:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List fines
      tags:
      - fines
    post:
      consumes:
      - application/json
      parameters:
      - description: Create fine request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateFineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create fine
      tags:
      - fines
  /fines/{id}:
    get:
      parameters:
      - description: Fine ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get fine
      tags:
      - fines
  /lendings:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List lendings
      tags:
      - lendings
    post:
      consumes:
      - application/json
      parameters:
      - description: Create lending request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateLendingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create lending
      tags:
      - lendings
  /lendings/{id}:
    delete:
      parameters:
      - description: Lending ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete lending
      tags:
      - lendings
    get:
      parameters:
      - description: Lending ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get lending
      tags:
      - lendings
    put:
      consumes:
      - application/json
      parameters:
      - description: Lending ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update lending request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateLendingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update lending
      tags:
      - lendings
  /librarian/hello:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Librarian hello
      tags:
      - access
  /reservations:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      parameters:
      - description: Create reservation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create reservation
      tags:
      - reservations
...
  /student/hello:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Student hello
      tags:
      - access
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"